requirements.txt
# Facial Recognition Attendance System Dependencies
# Python 3.10+ required

# Web framework
Flask>=3.0.0

# Face recognition library (requires dlib)
face-recognition>=1.3.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# For handling image uploads
Werkzeug>=3.0.0
